requirements.txt
pipecat-ai[daily,openai]
python-dotenv
aiohttp 